requirements.txt
Flask==2.0.1
opencv-python-headless==4.5.3.56
numpy==1.21.2
Pillow==8.3.1
scikit-learn==0.24.2
pandas==1.3.3
progressbar==2.5
mtcnn==0.1.0
tensorflow==2.6.0
keras==2.6.0


